Resolves: #8569

Introduces `shared` module, which is intendent to be used to store
components that are used across multiple modules and do not belong to
any particular one.

This is the first PR which should be used to discuss the details.

<!-- Use "x" to fill the checkboxes below like [x] -->

- [x] I signed [CLA](https://www.audacityteam.org/cla/)
- [x] The title of the pull request describes an issue it addresses
- [x] If changes are extensive, then there is a sequence of easily
reviewable commits
- [x] Each commit's message describes its purpose and effects
- [x] There are no behavior changes unnecessary for the stated purpose
of the PR

Recommended:
- [x] Each commit compiles and runs on my machine without known
undesirable changes of behavior

QA:

- [ ] Autobot test cases have been run
- [x] Smoke testing of the UI containing timecode input: main toolbar,
status bar, generators

+++ src/effects/builtin/dtmfgen/dtmfgenerator.cpp
@@ -173,7 +173,7 @@ size_t DtmfGenerator::Instance::ProcessBlock(
 
         if (isTone) {
             // generate the tone and append
-            assert(curSeqPos < dtmfSettings.dtmfNTones);
+            assert(curSeqPos < static_cast<int>(dtmfSettings.dtmfNTones));
             MakeDtmfTone(
                 buffer, len, mSampleRate, dtmfSettings.dtmfSequence[curSeqPos],
                 curTonePos, numSamplesTone, dtmfSettings.dtmfAmplitude);
